refactor: api regarcy 코드 정리
