+++ src/hooks/legacy/useRecruitingData.js
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react';
-import { googleFormsAPI, applicationsAPI } from '../../services/api';
+import { googleFormsAPI, applicationsAPI } from '../../services/api/index.js';
 import { PASS_STATUS_KOREAN } from '../../constants/recruitment';
 
 export const useRecruitingData = (id) => {
