+++ src/services/api/domains/admin/index.js
@@ -1,3 +1,4 @@
 // 관리자 도메인 API 통합 내보내기
 export { adminAPI } from './adminAPI.js';
 export { googleFormsAPI } from './googleFormsAPI.js';
+export { authAPI } from './authAPI.js';
