+++ src/services/api/domains/mail/index.js
@@ -1,0 +1,2 @@
+// 메일 도메인 API 통합 내보내기
+export { mailAPI } from './mailAPI.js';
